Removed plotting logic to separate module
